Update generic-helpers.js
I noticed when converting to word, it doesn't handle certain HTML entities well. It ends up displaying "&quot;" and so on.

Figured out by the time it reaches the wpml (which converts to WordProcessingML) function, the string was already escaped (although I don't know how it does that). So if you put quote (") in your resume JSON (e.g. a writing summary), it'll look like &quot; at the START of the wpml function.

The problem is in line 386, it uses xml-escape to escape it again. So it'll end up as "&amp;quot;", which in the end rendered on the document as "&quot;".

As such, I propose a change to ignore "&" when using xml-escape.